Added check for user authentication if User is interacting with SDK from inspector
Testing with guest account for now as user login is implemented in another branch. User can add two messages one for Authentication Success and Authentication Failed in the inspector after calling those methods in OnClick event of the Load button. Only if the authentication is successful, the presentation will load.